requirements.txt
Flask==3.0.3
fuzzywuzzy
sentence_transformers
transformers
torch
mmh3
pandas
numpy
langchain-anthropic
langchain-community
python-dotenv
pydantic